Extract performance horizon contracts
