+++ scripts/check_refactor_quality_thresholds.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import Iterable, Sequence
 
 DEFAULT_SOURCE_ROOT = Path("src/app")
-DEFAULT_MAX_SOURCE_FILE_LINES = 930
+DEFAULT_MAX_SOURCE_FILE_LINES = 914
 DEFAULT_MAX_FUNCTION_LINES = 49
